Quiz and Question services is running together

+++ Question_Service/src/main/java/com/arzezan/question_service/Controller/QuestionController.java
@@ -38,7 +38,7 @@ public class QuestionController {
     }
 
     @GetMapping("generate")
-    public ResponseEntity<List<Integer>> getQuestionsIdByCategory(@RequestParam String category, @RequestParam int amount) {
+    public ResponseEntity<List<Long>> getQuestionsIdByCategory(@RequestParam String category, @RequestParam int amount) {
         return questionService.getQuestionsIdByCategory(category, amount);
     }
